move usage dumping for readme to `mod dump_usage`
